Deprecating panos modules; pointing to Galaxy role
