+++ lib/ansible/modules/network/panos/_panos_dag.py
@@ -21,46 +21,42 @@
 
 DOCUMENTATION = '''
 ---
-module: panos_check
-short_description: check if PAN-OS device is ready for configuration
+module: panos_dag
+short_description: create a dynamic address group
 description:
-    - Check if PAN-OS device is ready for being configured (no pending jobs).
-    - The check could be done once or multiple times until the device is ready.
+    - Create a dynamic address group object in the firewall used for policy rules
 author: "Luigi Mori (@jtschichold), Ivan Bojer (@ivanbojer)"
 version_added: "2.3"
 requirements:
     - pan-python
+deprecated:
+    alternative: Use U(https://galaxy.ansible.com/PaloAltoNetworks/paloaltonetworks) instead.
+    removed_in: "2.11"
+    why: Consolidating code base.
 options:
-    timeout:
+    dag_name:
         description:
-            - timeout of API calls
-        required: false
-        default: "0"
-    interval:
+            - name of the dynamic address group
+        required: true
+    dag_filter:
         description:
-            - time waited between checks
-        required: false
-        default: "0"
+            - dynamic filter user by the dynamic address group
+        required: true
+    commit:
+        description:
+            - commit if changed
+        type: bool
+        default: 'yes'
 extends_documentation_fragment: panos
 '''
 
 EXAMPLES = '''
-# single check on 192.168.1.1 with credentials admin/admin
-- name: check if ready
-  panos_check:
+- name: dag
+  panos_dag:
     ip_address: "192.168.1.1"
     password: "admin"
-
-# check for 10 times, every 30 seconds, if device 192.168.1.1
-# is ready, using credentials admin/admin
-- name: wait for reboot
-  panos_check:
-    ip_address: "192.168.1.1"
-    password: "admin"
-  register: result
-  until: result is not failed
-  retries: 10
-  delay: 30
+    dag_name: "dag-1"
+    dag_filter: "'aws-tag.aws:cloudformation:logical-id.ServerInstance' and 'instanceState.running'"
 '''
 
 RETURN = '''
@@ -68,12 +64,11 @@ RETURN = '''
 '''
 
 ANSIBLE_METADATA = {'metadata_version': '1.1',
-                    'status': ['preview'],
+                    'status': ['deprecated'],
                     'supported_by': 'community'}
 
 
 from ansible.module_utils.basic import AnsibleModule
-import time
 
 try:
     import pan.xapi
@@ -81,17 +76,33 @@ try:
 except ImportError:
     HAS_LIB = False
 
+_ADDRGROUP_XPATH = "/config/devices/entry[@name='localhost.localdomain']" +\
+                   "/vsys/entry[@name='vsys1']/address-group/entry[@name='%s']"
 
-def check_jobs(jobs, module):
-    job_check = False
-    for j in jobs:
-        status = j.find('.//status')
-        if status is None:
-            return False
-        if status.text != 'FIN':
-            return False
-        job_check = True
-    return job_check
+
+def addressgroup_exists(xapi, group_name):
+    xapi.get(_ADDRGROUP_XPATH % group_name)
+    e = xapi.element_root.find('.//entry')
+    if e is None:
+        return False
+    return True
+
+
+def add_dag(xapi, dag_name, dag_filter):
+    if addressgroup_exists(xapi, dag_name):
+        return False
+
+    # setup the non encrypted part of the monitor
+    exml = []
+
+    exml.append('<dynamic>')
+    exml.append('<filter>%s</filter>' % dag_filter)
+    exml.append('</dynamic>')
+
+    exml = ''.join(exml)
+    xapi.set(xpath=_ADDRGROUP_XPATH % dag_name, element=exml)
+
+    return True
 
 
 def main():
@@ -99,8 +110,9 @@ def main():
         ip_address=dict(required=True),
         password=dict(required=True, no_log=True),
         username=dict(default='admin'),
-        timeout=dict(default=0, type='int'),
-        interval=dict(default=0, type='int')
+        dag_name=dict(required=True),
+        dag_filter=dict(required=True),
+        commit=dict(type='bool', default=True)
     )
     module = AnsibleModule(argument_spec=argument_spec, supports_check_mode=False)
     if not HAS_LIB:
@@ -109,33 +121,23 @@ def main():
     ip_address = module.params["ip_address"]
     password = module.params["password"]
     username = module.params['username']
-    timeout = module.params['timeout']
-    interval = module.params['interval']
 
     xapi = pan.xapi.PanXapi(
         hostname=ip_address,
         api_username=username,
-        api_password=password,
-        timeout=60
+        api_password=password
     )
 
-    checkpnt = time.time() + timeout
-    while True:
-        try:
-            xapi.op(cmd="show jobs all", cmd_xml=True)
-        except Exception:
-            pass
-        else:
-            jobs = xapi.element_root.findall('.//job')
-            if check_jobs(jobs, module):
-                module.exit_json(changed=True, msg="okey dokey")
+    dag_name = module.params['dag_name']
+    dag_filter = module.params['dag_filter']
+    commit = module.params['commit']
 
-        if time.time() > checkpnt:
-            break
+    changed = add_dag(xapi, dag_name, dag_filter)
 
-        time.sleep(interval)
+    if changed and commit:
+        xapi.commit(cmd="<commit></commit>", sync=True, interval=1)
 
-    module.fail_json(msg="Timeout")
+    module.exit_json(changed=changed, msg="okey dokey")
 
 
 if __name__ == '__main__':
